+++ Changelog.md
@@ -3,6 +3,9 @@
 1. Added SemanticVersion class
     * [Pull request 120](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/120)
 
+1. Added Matrix4<T>::LookAt
+    * [Pull request 124](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/124)
+
 1. Added triangle3 class, and helper functions in other classes
     * [Pull request 56](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/56)